+++ mettagrid/mettagrid/mettagrid_c.hpp
@@ -34,7 +34,7 @@ public:
 
   // Python API methods
   py::tuple reset();
-  py::tuple step(py::array_t<int> actions);
+  py::tuple step(py::array_t<long long> actions);
   void set_buffers(const py::array_t<unsigned char, py::array::c_style>& observations,
                    const py::array_t<bool, py::array::c_style>& terminals,
                    const py::array_t<bool, py::array::c_style>& truncations,
@@ -108,8 +108,8 @@ private:
                             unsigned short obs_width,
                             unsigned short obs_height,
                             size_t agent_idx);
-  void _compute_observations(py::array_t<int> actions);
-  void _step(py::array_t<int> actions);
+  void _compute_observations(py::array_t<long long> actions);
+  void _step(py::array_t<long long> actions);
 };
 
 #endif  // METTAGRID_METTAGRID_METTAGRID_C_HPP_
